+++ src/game/g_cmds.c
@@ -887,7 +887,7 @@ static qboolean G_SayTo( gentity_t *ent, gentity_t *other, saymode_t mode, const
   }
 
   trap_SendServerCommand( other-g_entities, va( "chat %d %d \"%s\"",
-    ent ? ent-g_entities : -1,
+    (int)( ent ? ent-g_entities : -1 ),
     mode,
     message ) );
 
@@ -915,7 +915,7 @@ void G_Say( gentity_t *ent, saymode_t mode, const char *chatText )
   {
     case SAY_ALL:
       G_LogPrintf( "Say: %d \"%s" S_COLOR_WHITE "\": " S_COLOR_GREEN "%s\n",
-        ( ent ) ? ent - g_entities : -1,
+        (int)( ( ent ) ? ent - g_entities : -1 ),
         ( ent ) ? ent->client->pers.netname : "console", chatText );
       break;
     case SAY_TEAM:
@@ -923,7 +923,7 @@ void G_Say( gentity_t *ent, saymode_t mode, const char *chatText )
       if( !ent || !ent->client )
         Com_Error( ERR_FATAL, "SAY_TEAM by non-client entity\n" );
       G_LogPrintf( "SayTeam: %d \"%s" S_COLOR_WHITE "\": " S_COLOR_CYAN "%s\n",
-        ent - g_entities, ent->client->pers.netname, chatText );
+        (int)( ent - g_entities ), ent->client->pers.netname, chatText );
       break;
     case SAY_RAW:
       if( ent )
@@ -968,7 +968,7 @@ static void Cmd_SayArea_f( gentity_t *ent )
     range[ i ] = g_sayAreaRange.value;
 
   G_LogPrintf( "SayArea: %d \"%s" S_COLOR_WHITE "\": " S_COLOR_BLUE "%s\n",
-    ent - g_entities, ent->client->pers.netname, msg );
+    (int)( ent - g_entities ), ent->client->pers.netname, msg );
 
   VectorAdd( ent->s.origin, range, maxs );
   VectorSubtract( ent->s.origin, range, mins );
@@ -1122,12 +1122,12 @@ void Cmd_VSay_f( gentity_t *ent )
     case VOICE_CHAN_LOCAL:
       trap_SendServerCommand( -1, va(
         "voice %d %d %d %d \"%s\"\n",
-        ent-g_entities, vchan, cmdNum, trackNum, text ) );
+        (int)( ent-g_entities ), vchan, cmdNum, trackNum, text ) );
       break;
     case VOICE_CHAN_TEAM:
       G_TeamCommand( ent->client->pers.teamSelection, va(
         "voice %d %d %d %d \"%s\"\n",
-        ent-g_entities, vchan, cmdNum, trackNum, text ) );
+        (int)( ent-g_entities ), vchan, cmdNum, trackNum, text ) );
       break;
     default:
       break;
@@ -1469,7 +1469,7 @@ void Cmd_CallVote_f( gentity_t *ent )
 
   G_LogPrintf( "%s: %d \"%s" S_COLOR_WHITE "\": %s\n",
     team == TEAM_NONE ? "CallVote" : "CallTeamVote",
-    ent - g_entities, ent->client->pers.netname, level.voteString[ team ] );
+    (int)( ent - g_entities ), ent->client->pers.netname, level.voteString[ team ] );
 
   if( team == TEAM_NONE )
   {
@@ -3338,7 +3338,7 @@ void Cmd_PrivateMessage_f( gentity_t *ent )
 
     G_LogPrintf( "%s: %d \"%s" S_COLOR_WHITE "\" \"%s\": ^%c%s\n",
       ( teamonly ) ? "TPrivMsg" : "PrivMsg",
-      ( ent ) ? ent - g_entities : -1,
+      (int)( ( ent ) ? ent - g_entities : -1 ),
       ( ent ) ? ent->client->pers.netname : "console",
       name, color, msg );
   }
