+++ src/game/g_main.c
@@ -1682,7 +1682,7 @@ void G_AdminMessage( gentity_t *ent, const char *msg )
   int     i;
 
   Com_sprintf( string, sizeof( string ), "chat %d %d \"%s\"",
-    ent ? ent - g_entities : -1,
+    (int)( ent ? ent - g_entities : -1 ),
     G_admin_permission( ent, ADMF_ADMINCHAT ) ? SAY_ADMINS : SAY_ADMINS_PUBLIC,
     msg );
 
@@ -1694,7 +1694,7 @@ void G_AdminMessage( gentity_t *ent, const char *msg )
   // Send to the logfile and server console
   G_LogPrintf( "%s: %d \"%s" S_COLOR_WHITE "\": " S_COLOR_MAGENTA "%s\n",
     G_admin_permission( ent, ADMF_ADMINCHAT ) ? "AdminMsg" : "AdminMsgPublic",
-    ent ? ent - g_entities : -1, ent ? ent->client->pers.netname : "console",
+    (int)( ent ? ent - g_entities : -1 ), ent ? ent->client->pers.netname : "console",
     msg );
 }
